Stock-Tracker/requirements.txt
requests>=2.31.0
schedule>=1.2.0
finnhub-python>=2.4.18
anthropic>=0.25.0
tzdata>=2024.1
